+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+All notable changes to this project will be documented in this file.
+
+
+
+## 0.0.2
+- Created the `global` package.
+- Integrated the `GetIt` service locator to allow for easy service registration and retrieval.
+- Added the `sl` singleton instance for global access to registered services.
+
+### Example usage:
+- Registered a sample service `MyService` to demonstrate how to use the package.
+
+
+## 0.0.1
+- Initial release